requirements.txt
torch
torchvision
flask
numpy
opencv-python-headless
Pillow
